requirements.txt
click
docker-py
pyyaml
sqlalchemy

#flask

#python-nginx fails import on python3